requirements.txt
flask
python-dotenv
pypdf==4.3.1
langchain-core
langchain-google-genai
google-generativeai
json5
gunicorn
flask-cors
